+++ src/postgres_mcp/database_health/index_health_calc.py
@@ -1,8 +1,7 @@
 from typing import Any
 
-from postgres_mcp.dta.safe_sql import SafeSqlDriver
-
-from ..dta.sql_driver import SqlDriver
+from ..sql import SafeSqlDriver
+from ..sql import SqlDriver
 
 
 class IndexHealthCalc:
