+++ src/postgres_mcp/database_health/vacuum_health_calc.py
@@ -1,8 +1,8 @@
 from dataclasses import dataclass
 
-from postgres_mcp.dta.safe_sql import SafeSqlDriver
+from ..sql import SafeSqlDriver
 
-from ..dta.sql_driver import SqlDriver
+from ..sql import SqlDriver
 
 
 @dataclass
