+++ src/postgres_mcp/dta/__init__.py
@@ -1,1 +1,23 @@
-"""PostgreSQL Database Tuning Advisor (DTA) tool."""
+from .dta_calc import DatabaseTuningAdvisor
+from .dta_calc import DTASession
+from .dta_calc import ColumnCollector
+from .dta_calc import ConditionColumnCollector
+from .dta_calc import Index
+from .dta_calc import IndexConfig
+from pglast import parse_sql
+from .dta_tools import DTATool
+from .dta_calc import MAX_NUM_DTA_QUERIES_LIMIT
+from .artifacts import ExplainPlanArtifact
+
+__all__ = [
+    "DatabaseTuningAdvisor",
+    "DTASession",
+    "ColumnCollector",
+    "ConditionColumnCollector",
+    "Index",
+    "IndexConfig",
+    "parse_sql",
+    "DTATool",
+    "MAX_NUM_DTA_QUERIES_LIMIT",
+    "ExplainPlanArtifact",
+]
