+++ src/postgres_mcp/sql/__init__.py
@@ -1,0 +1,6 @@
+"""SQL utilities."""
+
+from .safe_sql import SafeSqlDriver
+from .sql_driver import DbConnPool, obfuscate_password, SqlDriver
+
+__all__ = ["SafeSqlDriver", "SqlDriver", "DbConnPool", "obfuscate_password"]
